# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  customer: 'customer',
  worker: 'worker',
  cooperative_admin: 'cooperative_admin',
} as const;

export interface Session {
  userId: string;
  name: string;
  email: string;
  role: Role;
  cooperativeId: string;
}

export type MetricTone = typeof MetricTone[keyof typeof MetricTone];


export const MetricTone = {
  blue: 'blue',
  green: 'green',
  orange: 'orange',
  purple: 'purple',
  red: 'red',
} as const;

export interface Metric {
  label: string;
  value: string;
  detail: string;
  tone: MetricTone;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: string;
  customerName: string;
  workerName: string;
  service: string;
  date: string;
  time: string;
  address: string;
  description: string;
  status: BookingStatus;
  amount: number;
}

export interface DashboardSummary {
  role: Role;
  metrics: Metric[];
  recentBookings: Booking[];
}

export interface BookingInput {
  workerId: string;
  service: string;
  date: string;
  time: string;
  address: string;
  description: string;
}

export type BookingStatusInputStatus = typeof BookingStatusInputStatus[keyof typeof BookingStatusInputStatus];


export const BookingStatusInputStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusInput {
  status: BookingStatusInputStatus;
}

export type WorkerAvailability = typeof WorkerAvailability[keyof typeof WorkerAvailability];


export const WorkerAvailability = {
  Available: 'Available',
  Busy: 'Busy',
  Unavailable: 'Unavailable',
} as const;

export interface Worker {
  id: string;
  name: string;
  skill: string;
  rating: number;
  experience: string;
  availability: WorkerAvailability;
  verified: boolean;
  location: string;
  jobsCompleted: number;
  initials: string;
}

export interface Availability {
  available: boolean;
  days: string[];
  startTime: string;
  endTime: string;
}

export interface AvailabilityInput {
  available: boolean;
  days: string[];
  startTime: string;
  endTime: string;
}

export interface Review {
  id: string;
  reviewer: string;
  subject: string;
  rating: number;
  comment: string;
  date: string;
}

export type ComplaintStatus = typeof ComplaintStatus[keyof typeof ComplaintStatus];


export const ComplaintStatus = {
  open: 'open',
  in_review: 'in_review',
  resolved: 'resolved',
} as const;

export interface Complaint {
  id: string;
  subject: string;
  description: string;
  status: ComplaintStatus;
  createdAt: string;
}

export interface ComplaintInput {
  subject: string;
  description: string;
}

export interface WorkerVerificationInput {
  verified: boolean;
  enabled: boolean;
}

export type RoleParameter = Role;

export type StatusParameter = string | null;

export type GetDashboardSummaryParams = {
role?: RoleParameter;
};

export type ListBookingsParams = {
role?: RoleParameter;
/**
 * @nullable
 */
status?: StatusParameter;
};

export type ListWorkersParams = {
/**
 * @nullable
 */
service?: string | null;
/**
 * @nullable
 */
availability?: string | null;
/**
 * @nullable
 */
minRating?: number | null;
};