# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: string;
  customerName: string;
  workerName: string;
  service: string;
  date: string;
  time: string;
  address: string;
  description: string;
  status: BookingStatus;
  amount: number;
}