# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  workerId: string;
  service: string;
  date: string;
  time: string;
  address: string;
  description: string;
}