# lib/api-zod/src/generated/types/bookingStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatusInputStatus } from './bookingStatusInputStatus';

export interface BookingStatusInput {
  status: BookingStatusInputStatus;
}