# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatus } from './complaintStatus';

export interface Complaint {
  id: string;
  subject: string;
  description: string;
  status: ComplaintStatus;
  createdAt: string;
}