# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintInput {
  subject: string;
  description: string;
}