# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from './booking';
import type { Metric } from './metric';
import type { Role } from './role';

export interface DashboardSummary {
  role: Role;
  metrics: Metric[];
  recentBookings: Booking[];
}