# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */
import type { MetricTone } from './metricTone';

export interface Metric {
  label: string;
  value: string;
  detail: string;
  tone: MetricTone;
}