# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  reviewer: string;
  subject: string;
  rating: number;
  comment: string;
  date: string;
}