# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  customer: 'customer',
  worker: 'worker',
  cooperative_admin: 'cooperative_admin',
} as const;