# lib/api-zod/src/generated/types/workerVerificationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CoopWork role-based cooperative services API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerVerificationInput {
  verified: boolean;
  enabled: boolean;
}